Split exercises module into focused files
Replace the monolithic packages/contents/_lib/exercises.ts with four new
modules: collection, renderable, set, and source. Update import sites
across
apps to use the new paths and adjust references to getExerciseCount,
getExercisesContent, getExerciseByNumber, and related helpers. Remove
the
ChoicesValidationError from shared errors.
